Fix bug with infinite OffsetBezier recursion.
Due to numeric precision, with very close bezier points,
OffsetBezier would fail to converge. Now we check how close
the new midpoint is when we recurse, and stop if the midpoint
ends up the same as one of the endpoints.

Also, beziers that curve straight back on themselves are better
handled (yeah, should be separate commit, I know) and
beziers that are just the same point repeated (or close to it)
are ignored more simply.