Change rasterizer output.
The rasterizer now outputs CoverageSpans, which are like the old EdgeSpans,
except the 'l' value is now the length of the span with coverage 'c'. This is simpler
and makes more sense than the old spans, which had an implied starting pixel
and an implied coverage value of 100% for the span.

Also, PixelCoverages are now called CoveragePoints.