Fix bleeding stroke bug.
If the pen path contained a Bezier with very close coordinates
(yet not exactly the same), we used to produce poor outlines that
would not rasterize well (bleeding, due to disconnected outlines).
This was due to floating point precision issues. We now check for this
type of Bezier in SanitizeBezierPath and do not forward it along
in the stream.

Ideally, we would repair the stream to make sure that the sanitized
path was fully connected despite the removal of the micro-Beziers.
In practice, the exclusion of the micro-Beziers is unlikely to make a
difference. But if needed, SanitizeBezierPath could be changed
to hold onto the most recent valid Bezier, and just adjust the endpoint
to be the endpoint of the micro-Bezier being removed.

I have prototyped this approach, and it works well, but I left it out
because it significantly complicates SanitizeBezierPath for very little
gain (at the moment). In addition, with this approach, it is
still possible to turn the valid Bezier into a micro-Bezier by "walking" the
endpoint 'C' over to 'A' or 'B' through many micro-Beziers. Very unlikely, though.

For now, the approach in this commit seems to work.