Implement new image structure with associated gate.
Now, images have their own public structure (gezira_Image_t) with init()
and reset_gate() functions. WriteToImage, ReadFromImage and Composite...Image
now take image structs as the argument, and the gating of these processes
is taken care of inside the contructors of these processes.

gezira_Image_reset_gate is meant to be called whenever the image will be
used in a different Nile runtime instance. This is because the gate associated with the
image is not valid in the new runtime, and thus we must reset the gate.

ReadFromImage takes an argument "skipNextGate" which, if non-zero, will
prevent the creation of a new gate for the ReadFromImage to control.
This is useful to activate on pipelines that could experience deadlock
if a process in the pipeline gates a later process, such as a ReadFromImage
that gates a later WriteToImage on the same image. The gate is
not needed in this case, thus we skip the gating between the two.